* Moved core.__init__ to core.py
* Moved everything out of the core module

* core.base_commands and core.settings both moved to private "_core" module

* core.downloads, core.cvar, and core.excepthook moved to downloads, cvars, and excepthooks

* Made fixes according to the core changes

* Minor fix to events.variable.LongVariable

* Fixed missing imports in some message types modules

* Added translation strings to sp command module, addons module, auth module, and excepthooks module

* Fixed error when using TranslationStrings instance with <CvarManager instance>.text()

+++ addons/source-python/_libs/events/variable.py
@@ -63,7 +63,7 @@ class ShortVariable(_EventVariable):
 
 class LongVariable(_EventVariable):
     '''Class used to store long event variables'''
-    _method = 'float'
+    _method = 'int'
 
 
 class FloatVariable(_EventVariable):
